src/requirements.txt
numpy
PyOpenGL-accelerate
PyOpenGL